refactor: auth 도메인 flat 구조 재설계 및 외부 auth 코드 반입
- signup/ 하위 폴더 해체 → auth.api/hooks/types/schemas/constants/lib.ts로 병합
- index.ts 신규: 외부 공개 심볼만 named export (export * 제거)
- 로그인 모달 클러스터를 shared에서 auth로 반입
  (kakao-login-button, use-login-modal, login-modal-query, open-login-modal-on-page)
- 카카오 콜백 route 도메인 로직 추출, route.ts는 HTTP 경계만 잔류
- deep import 전부 @/features/auth 배럴 경유로 교체
