+++ features/auth/components/signup-flow.tsx
@@ -6,14 +6,14 @@ import { FormProvider } from 'react-hook-form';
 
 import { StepIndicator } from '@/shared/components/step-indicator';
 
-import type { SignupStepId } from '../signup.constants';
-import type { SignupFormData } from '../types';
+import type { SignupStepId } from '../auth.constants';
+import type { SignupFormData } from '../auth.types';
 
 import { SignupStepBasic } from './signup-step-basic';
 import { SignupStepCareer } from './signup-step-career';
 import { SignupStepComplete } from './signup-step-complete';
-import { useSignupForm } from '../hooks/use-signup-form';
-import { SIGNUP_STEPS, SIGNUP_STEP_DESCRIPTIONS } from '../signup.constants';
+import { SIGNUP_STEPS, SIGNUP_STEP_DESCRIPTIONS } from '../auth.constants';
+import { useSignupForm } from '../auth.hooks';
 
 const STEP_BASIC_FIELDS = [
   'name',
