--- features/auth/signup/hooks/use-signup.mutation.ts
@@ -1,9 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-
-import { createSignup } from '../api/create-signup.api';
-
-export const useSignup = () => {
-  return useMutation({
-    mutationFn: createSignup,
-  });
-};
