--- features/auth/types.ts
@@ -1,4 +1,0 @@
-export interface LoginResponse {
-  accessToken: string;
-  userId: number;
-}
