--- shared/constants/login-modal-query.ts
@@ -1,2 +1,0 @@
-export const LOGIN_MODAL_QUERY_KEY = 'modal';
-export const LOGIN_MODAL_QUERY_VALUE = 'login';
